DelegateAction struct, split receipts file
