+++ lake-primitives/src/lib.rs
@@ -2,7 +2,7 @@ pub use near_indexer_primitives::{
     self, near_primitives, types::AccountId, CryptoHash, IndexerShard, StreamerMessage,
 };
 
-pub use types::{block, events, receipts, state_changes, transactions};
+pub use types::{actions, block, events, receipts, state_changes, transactions};
 
 mod types;
