+++ lake-primitives/src/types/block.rs
@@ -1,7 +1,8 @@
 use std::collections::HashMap;
 
+use super::actions::{self, ActionMetaDataExt};
 use super::events::{self, EventsTrait};
-use super::receipts::{self, ActionMetaDataExt};
+use super::receipts::{self};
 use super::state_changes;
 use super::transactions;
 use crate::near_indexer_primitives::{types::AccountId, views, CryptoHash, StreamerMessage};
@@ -12,7 +13,7 @@ pub struct Block {
     executed_receipts: Vec<receipts::Receipt>,
     postponed_receipts: Vec<receipts::Receipt>,
     transactions: Vec<transactions::Transaction>,
-    actions: Vec<receipts::Action>,
+    actions: Vec<actions::Action>,
     events: HashMap<super::ReceiptId, Vec<events::Event>>,
     state_changes: Vec<state_changes::StateChange>,
 }
@@ -86,19 +87,19 @@ impl Block {
         self.transactions.iter()
     }
 
-    fn actions_from_streamer_message(&self) -> Vec<receipts::Action> {
+    fn actions_from_streamer_message(&self) -> Vec<actions::Action> {
         self.streamer_message()
             .shards
             .iter()
             .flat_map(|shard| shard.receipt_execution_outcomes.iter())
             .filter_map(|receipt_execution_outcome| {
-                receipts::Action::try_vec_from_receipt_view(&receipt_execution_outcome.receipt).ok()
+                actions::Action::try_vec_from_receipt_view(&receipt_execution_outcome.receipt).ok()
             })
             .flatten()
             .collect()
     }
 
-    pub fn actions(&mut self) -> impl Iterator<Item = &receipts::Action> {
+    pub fn actions(&mut self) -> impl Iterator<Item = &actions::Action> {
         if self.actions.is_empty() {
             self.build_actions_cache();
         }
@@ -128,7 +129,7 @@ impl Block {
     pub fn actions_by_receipt_id<'a>(
         &'a mut self,
         receipt_id: &'a super::ReceiptId,
-    ) -> impl Iterator<Item = &receipts::Action> + 'a {
+    ) -> impl Iterator<Item = &actions::Action> + 'a {
         self.actions()
             .filter(move |action| &action.receipt_id() == receipt_id)
     }
