+++ lake-primitives/src/types/mod.rs
@@ -1,5 +1,8 @@
+pub mod actions;
 pub mod block;
+mod delegate_actions;
 pub mod events;
+mod impl_actions;
 pub mod receipts;
 pub mod state_changes;
 pub mod transactions;
